MPStaticAnalysis/requirements.txt
argparse==1.4.0
pyyaml
beautifulsoup4==4.9.1
bs4==0.0.1
EditorConfig==0.12.2
esprima==4.0.1
jsbeautifier==1.13.0
neobolt==1.7.17
neomodel==4.0.1
neotime==1.7.4
pyotp==2.3.0
python-dotenv==0.14.0
requests==2.24.0
selenium==3.141.0
urllib3==1.25.10
pandas==1.1.5
pyvirtualdisplay
tldextract